+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Breaking changes
 
 * The decimal type must now implement the `beancount_parser_2::Decimal` *sealed* trait.
+  At the moment, only `f64` and `rust_decimal::Decimal` implements it. The later requires the `rust_decimal` feature flag.
+
+
+### Added
+
+* `rust_decimal` feature flag providing an implementation of `beancount_parser_2::Decimal` for `rust_decimal::Decimal`
 
 
 ## [1.0.0-alpha.2] - 2023-05-22