+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### New syntax supported
 
-* Amount value expressions
+* Expression for amount value
 
 
 ### Documentation